Subindo correcoes

+++ src/main/java/com/restaurante/entity/Mesa.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -16,5 +17,5 @@ import java.util.List;
 public class Mesa {
     private Integer id;
     private List<Pedido> listaPedidos;
-    private Integer valorTotalConsumido;
+    private BigDecimal valorTotalConsumido;
 }
